docker: Remove creation of `/opt/openvidu-call/.env` in entrypoint.sh
If `.env` file is not present, the backend will simply use the
defined environment variables. It is not necessary to create an
environment file in the entrypoint script.

Also, by removing the creation of the `.env` file in the entrypoint,
the user can run the container with its own `.env` file using
`-v /path/to/.env:/opt/openvidu-call/.env` in the `docker run`.